chore: reorg component structure a bit
