+++ solution/ui/prototype/src/components/floating_nav/BottomNavBtn.vue
